Including product description in functions

+++ db/admin/products.js
@@ -10,11 +10,16 @@ const createProduct = async(req, res) => {
     const id = idGen(5);
 
     // VALIDATION AND SANITISATION
-    let { name, price, category } = req.body;
+    let { name, description, price, category } = req.body;
 
     // Name
     if (typeof name !== "string") return res.status(400).send("Error: Name must be a string.");
     name = sanitizeHtml(trim(escape(name)));
+
+    // Description
+    if (typeof description !== "string") return res.status(400).send("Error: Description must be a string.");
+    description = sanitizeHtml(trim(escape(description)));
+    if (!isLength(description, { min: 1, max: 255 })) return res.status(400).send("Error: Description must be no longer than 255 characters.");
 
     // Price
     if (typeof price !== "string" && typeof price !== "number") return res.status(400).send("Error: Price must be a number.");
@@ -25,14 +30,15 @@ const createProduct = async(req, res) => {
     if (typeof category !== "string") return res.status(400).send("Error: Category must be a string.");
     category = sanitizeHtml(trim(escape(category))).toLowerCase();
 
-    try {
-        // Send error if product already exists
+    try { // Get product
         let result = await pool.query("SELECT * FROM products WHERE name = $1", [name]);
+
+        // Send error if product already exists
         if (result.rows.length > 0) return res.status(409).send("Error: A product with the provided name already exists.");
 
         // Create product
-        let text = `INSERT INTO products (id, name, price, category, created_at) VALUES ($1, $2, $3, $4, to_timestamp(${Date.now()} / 1000)) RETURNING id`;
-        let values = [id, name, price, category];
+        let text = `INSERT INTO products (id, name, description, price, category, created_at) VALUES ($1, $2, $3, $4, $5, to_timestamp(${Date.now()} / 1000)) RETURNING id`;
+        let values = [id, name, description, price, category];
         result = await pool.query(text, values);
         res.status(201).send(`Product created with ID: ${result.rows[0].id}`);
     } catch (err) {
@@ -45,13 +51,16 @@ const updateProduct = async(req, res) => {
     if (!req.query.id) return res.status(400).send("Error: No product ID provided.");
 
     // VALIDATION
-    let { name, price, category } = req.body;
+    let { name, description, price, category } = req.body;
 
     // Send error if no details are provided
-    if (!name && !price && !category) return res.status(400).send("Error: No updates provided.");
+    if (!name && !description && !price && !category) return res.status(400).send("Error: No updates provided.");
 
     // Name
     if (name && typeof name !== "string") return res.status(400).send("Error: Name must be a string.");
+
+    // Description
+    if (description && typeof description !== "string") return res.status(400).send("Error: Description must be a string.");
 
     // Price
     if (price && typeof price !== "string" && typeof price !== "number") return res.status(400).send("Error: Price must be a number.");
@@ -63,29 +72,33 @@ const updateProduct = async(req, res) => {
     let id = trim(req.query.id);
     if (!isNumeric(id, { no_symbols: true }) || !isLength(id, { min: 5, max: 5 })) return res.status(400).send("Error: Invalid product ID provided.");
 
-    try { // Get product details
-        let result = await pool.query("SELECT name, price, category FROM products WHERE id = $1", [id]);
+    try { // Get product
+        let result = await pool.query("SELECT name, description, price, category FROM products WHERE id = $1", [id]);
 
         // Send error if product does not exist
         if (result.rows.length === 0) return res.status(404).send("Error: This product does not exist.");
 
         // Save existing details to object
-        let old = { name: result.rows[0].name, price: result.rows[0].price, category: result.rows[0].category };
+        let old = {...result.rows[0] };
 
         // SANITISATION
         // Name
-        name = sanitizeHtml(trim(escape(name || result.rows[0].name)));
+        name = sanitizeHtml(trim(escape(name || old.name)));
+
+        // Description
+        description = sanitizeHtml(trim(escape(description || old.description)));
+        if (!isLength(description, { min: 1, max: 255 })) return res.status(400).send("Error: Description must be no longer than 255 characters.");
 
         // Price
-        price = price || result.rows[0].price;
+        price = price || old.price;
         if (typeof price === "string") price = trim(price);
         price = Math.round(price);
 
         // Category
-        category = sanitizeHtml(trim(escape(category || result.rows[0].category)));
+        category = sanitizeHtml(trim(escape(category || old.category)));
 
         // Send error if no updates made
-        if (old.name === name && old.price === price && old.category === category) return res.status(400).send("Error: No updates provided.");
+        if (old.name === name && old.description === description && old.price === price && old.category === category) return res.status(400).send("Error: No updates provided.");
 
         // Send error if another product of the desired name already exists
         if (req.body.name) {
@@ -93,8 +106,8 @@ const updateProduct = async(req, res) => {
             if (result.rows.length > 0) return res.status(409).send("Error: A product with the provided name already exists.");
         }
 
-        // Update product details
-        result = await pool.query("UPDATE products SET name = $1, price = $2, category = $3 WHERE id = $4 RETURNING id", [name, price, category, id]);
+        // Update product
+        result = await pool.query("UPDATE products SET name = $1, description = $2, price = $3, category = $4 WHERE id = $5 RETURNING id", [name, description, price, category, id]);
         res.status(200).send(`Product updated with ID: ${result.rows[0].id}`);
     } catch (err) {
         res.status(500).send("An unknown error occurred. Kindly try again.");
@@ -109,8 +122,14 @@ const deleteProduct = async(req, res) => {
     let id = trim(req.query.id);
     if (!isNumeric(id, { no_symbols: true }) || !isLength(id, { min: 5, max: 5 })) return res.status(400).send("Error: Invalid product ID provided.");
 
-    try { // Delete product
-        let result = await pool.query("DELETE FROM products WHERE id = $1", [id]);
+    try { // Get product
+        let result = await pool.query("SELECT * FROM products WHERE id = $1", [id]);
+
+        // Send error if product does not exist
+        if (result.rows.length === 0) return res.status(404).send("Error: This product does not exist.");
+
+        // Delete product
+        result = await pool.query("DELETE FROM products WHERE id = $1", [id]);
         res.status(204).send("Product deleted successfully");
     } catch (err) {
         res.status(500).send("An unknown error occurred. Kindly try again.");
